+++ src/main/java/kitchenpos/eatinorders/tobe/domain/order/createsupporter/OrderedMenuReader.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain.order.create_support;
+package kitchenpos.eatinorders.tobe.domain.order.createsupporter;
 
 import java.util.List;
 import java.util.UUID;
