+++ src/main/java/kitchenpos/eatinorders/tobe/domain/ordertable/OrderTableName.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain.order_table;
+package kitchenpos.eatinorders.tobe.domain.ordertable;
 
 import java.util.Objects;
