+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/EatInOrderCreatorTest.java
@@ -12,10 +12,10 @@ import java.util.NoSuchElementException;
 import java.util.UUID;
 import kitchenpos.eatinorders.tobe.domain.fake.FakeOrderTableReader;
 import kitchenpos.eatinorders.tobe.domain.fake.FakeOrderedMenuReader;
-import kitchenpos.eatinorders.tobe.domain.order.create_support.OrderTable;
-import kitchenpos.eatinorders.tobe.domain.order.create_support.OrderTableReader;
-import kitchenpos.eatinorders.tobe.domain.order.create_support.OrderedMenu;
-import kitchenpos.eatinorders.tobe.domain.order.create_support.OrderedMenuReader;
+import kitchenpos.eatinorders.tobe.domain.order.createsupporter.OrderTable;
+import kitchenpos.eatinorders.tobe.domain.order.createsupporter.OrderTableReader;
+import kitchenpos.eatinorders.tobe.domain.order.createsupporter.OrderedMenu;
+import kitchenpos.eatinorders.tobe.domain.order.createsupporter.OrderedMenuReader;
 import kitchenpos.eatinorders.tobe.domain.order.dto.OrderCreateCommand;
 import kitchenpos.eatinorders.tobe.domain.order.dto.OrderLineItemCreateCommand;
 import kitchenpos.supports.domain.tobe.OrderType;
